requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pymupdf==1.23.8
pillow==10.2.0
anthropic
aiofiles==23.2.1
jinja2==3.1.3
reportlab==4.0.8
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0